Added IRCBot::get_nickname() and IRCBot::get_channel().
This was done so that I could then add IRCCommandHandler::is_channel_msg(), which checks for whether it is a PRIVMSG for the channel the bot is in.

And that was done so I could change UserStats to only update the word count when a message is for the channel, so private messages are not included.



git-svn-id: svn+ssh://white/home/lajnold/svn/IRCBot@28 712a1f97-fb38-4509-82a9-c8a0c282b14c