Finally something I have wanted to do a long time...
Instead of storing the various CommandHandlers as separate objects in IRCBot, store them as a list of 'CommandHandler *'s. The command handlers can now be created dynamically in IRCBot::IRCBot, and called polymorphically in IRCBot::parse_IRC_message.

This allows loosening some of the dependencies between IRCBot.h and the various handler headers.



git-svn-id: svn+ssh://white/home/lajnold/svn/IRCBot@80 712a1f97-fb38-4509-82a9-c8a0c282b14c