A fairly substantial rewrite of various bits.
There is a new Message class. It does some basic dissecting of messages, with methods for accessing the parts. This dissecting previously resided in CommandHandler. CommandHandler still does the dissecting specific to the '!' commands. The command handlers get passed a Message instead of the previous raw message string. The raw string can still be accessed through the Message.

One can now get both the nick and the hostname of the sender of a message.

The packet_t typedef has been renamed to StringList. It is still a std::vector<std::string>.

There are some new generic utility functions in utils.h (utils.cpp is gone):
- split_whitespace
- find_nth
- contains
- remove_val
- remove_pred

And probably some more changes.


git-svn-id: svn+ssh://white/home/lajnold/svn/IRCBot@122 712a1f97-fb38-4509-82a9-c8a0c282b14c